+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/greet/HelloData.scala
@@ -1,0 +1,20 @@
+package com.github.kkrull.http4s.greet
+
+import io.circe.{Encoder, Json}
+import org.http4s.EntityEncoder
+import org.http4s.circe._
+
+final case class Name(name: String) extends AnyVal
+
+final case class Greeting(greeting: String) extends AnyVal
+
+object Greeting {
+  implicit val greetingEncoder: Encoder[Greeting] = new Encoder[Greeting] {
+    final def apply(a: Greeting): Json = Json.obj(
+      ("message", Json.fromString(a.greeting)),
+    )
+  }
+
+  implicit def greetingEntityEncoder[F[_]]: EntityEncoder[F, Greeting] =
+    jsonEncoderOf[F, Greeting]
+}
