+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/greet/HelloRoutes.scala
@@ -2,7 +2,6 @@ package com.github.kkrull.http4s.greet
 
 import cats.effect.Sync
 import cats.implicits._
-import com.github.kkrull.http4s.joke.Jokes
 import org.http4s.HttpRoutes
 import org.http4s.dsl.Http4sDsl
 
@@ -13,20 +12,8 @@ object HelloRoutes {
     import dsl._
     HttpRoutes.of[F] { case GET -> Root / "hello" / name =>
       for {
-        greeting <- H.greet(HelloWorld.Name(name))
+        greeting <- H.greet(Name(name))
         resp <- Ok(greeting)
-      } yield resp
-    }
-  }
-
-  def jokeRoutes[F[_]: Sync](J: Jokes[F]): HttpRoutes[F] = {
-    val dsl = new Http4sDsl[F] {}
-
-    import dsl._
-    HttpRoutes.of[F] { case GET -> Root / "joke" =>
-      for {
-        joke <- J.get
-        resp <- Ok(joke)
       } yield resp
     }
   }
