+++ http4s-quickstart/src/main/scala/com/github/kkrull/http4s/greet/HelloWorld.scala
@@ -2,35 +2,14 @@ package com.github.kkrull.http4s.greet
 
 import cats.Applicative
 import cats.implicits._
-import io.circe.{Encoder, Json}
-import org.http4s.EntityEncoder
-import org.http4s.circe._
 
 trait HelloWorld[F[_]] {
-  def greet(n: HelloWorld.Name): F[HelloWorld.Greeting]
+  def greet(n: Name): F[Greeting]
 }
 
 object HelloWorld {
-  final case class Name(name: String) extends AnyVal
-
-  /** More generally you will want to decouple your edge representations from your internal data structures, however
-    * this shows how you can create encoders for your data.
-    */
-  final case class Greeting(greeting: String) extends AnyVal
-
-  object Greeting {
-    implicit val greetingEncoder: Encoder[Greeting] = new Encoder[Greeting] {
-      final def apply(a: Greeting): Json = Json.obj(
-        ("message", Json.fromString(a.greeting)),
-      )
-    }
-
-    implicit def greetingEntityEncoder[F[_]]: EntityEncoder[F, Greeting] =
-      jsonEncoderOf[F, Greeting]
-  }
-
   def impl[F[_]: Applicative]: HelloWorld[F] = new HelloWorld[F] {
-    def greet(n: HelloWorld.Name): F[HelloWorld.Greeting] =
+    def greet(n: Name): F[Greeting] =
       Greeting("Hello, " + n.name).pure[F]
   }
 }
